backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
firebase-admin==6.5.0
qiskit==1.1.2
# qiskit-aer requires Visual Studio Build Tools on Windows
# Install with: pip install qiskit-aer
# Or install Visual Studio Build Tools from: https://visualstudio.microsoft.com/downloads/
# qiskit-aer==0.14.2
google-generativeai==0.8.3
supabase==2.5.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
aiofiles==24.1.0

